Created separate file for fetch status functions. Cleaned up imports.
